+++ src/Shgif/Type.hs
@@ -6,6 +6,8 @@ module Shgif.Type (
     Format(..), Shgif(..)
     , shgifToCanvas, updateShgifNoLoop, updateShgif, updateShgifReversedNoLoop, updateShgifReversed, getShgif, getShgifs
     , updateShgifTo
+    , fromCanvas
+    , fromTartFile
     , canvas, width, height
 ) where
 
@@ -21,7 +23,14 @@ import Data.Yaml (FromJSON(..), withObject, (.:), Object(..), withArray
                  , withText
                  , Parser(..), Value(..), ParseException
                  , decodeFileEither)
-import Tart.Canvas (Canvas, canvasFromText, newCanvas)
+import Tart.Canvas (Canvas, canvasFromText, newCanvas, canvasSize, prettyPrintCanvas)
+import Tart.Format (TartFile, sortedCanvases, tartFileCanvasOrder, tartFileCanvasList)
+
+import Control.Arrow (second)
+
+-- Config {{{
+defaultTimeStampInterval = 100
+--- }}}
 
 -- | Format  of shgif file
 --
@@ -135,6 +144,43 @@ getShgifs xs = do
     fromLeft (Left e)   = e
     fromRight (Right a) = a
     caughtExceptions rs = map fromLeft $ filter isLeft rs
+
+
+-- | Create 'Shgif' from 'Tart.Canvas.Canvas'es without any meta value
+--
+-- If you can implement meta values, use 'fromCanvasWithMeta' instead.
+--
+-- The first argument specify the tick for each frame.
+--
+-- If 'Nothing', use 'defaultTimeStampInterval'
+fromCanvas :: Maybe [Int] -> [Canvas] -> Shgif
+fromCanvas (Just ts)  cs = fromCanvasWithMeta "" "" ts cs
+fromCanvas Nothing cs = fromCanvasWithMeta "" "" defaultTimestamps cs
+    where
+        defaultTimestamps = take (length cs) $ 0: interval 0 defaultTimeStampInterval
+        interval orig i = orig + i: interval (orig + i) i
+
+-- | Create 'Shgif' from 'Tart.Format.TartFile'
+fromTartFile :: Maybe [Int] -> TartFile -> Shgif
+fromTartFile ix tartFile = let cs  = sortedCanvases (tartFileCanvasOrder tartFile) (tartFileCanvasList tartFile)
+                           in        fromCanvas Nothing $ cs ++ [last cs]
+
+
+-- | Create 'Shgif' from 'Tart.Canvas.Canvas' with meta value
+--
+-- This only support 'Page' format, because 'Tart.Canvas.Canvas' is Bitmap image.
+--
+-- For __Not__ 'updateShgifTo' use cases:
+-- Make sure to __duplicate the last Canvas__ so that it'll show up for more than 1 tick.
+--
+-- (If you don't, The last frame only appear for 1 tick. In most case, it's the same as invisible)
+fromCanvasWithMeta :: String -> String -> [Int] -> [Canvas] -> Shgif
+fromCanvasWithMeta title author timestamps cs = Shgif title author Page w h 0 convertedData Nothing
+  where
+    (w, h) = foldr1 (\(x,y ) (x', y') -> (max x x', max y y')) whList
+    whList = fmap (canvasSize) cs
+    convertedData :: [TimeStamp]
+    convertedData = zip timestamps $ fmap (lines . prettyPrintCanvas False . pure) cs
 
 
 -- | Update 'Shgif''s internal tick state, which will affect frame rendering.